update drivetime band client and configuration namespacing
